+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [3.7.0] - 2020-12-02
+
 ### Added
 
 - Validate App CR configmap and secret names if managed by config-controller.
@@ -109,7 +112,8 @@ since there can be a delay creating them on cluster creation.
 
 - First release
 
-[Unreleased]: https://github.com/giantswarm/app/compare/v3.6.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app/compare/v3.7.0...HEAD
+[3.7.0]: https://github.com/giantswarm/app/compare/v3.5.0...v3.7.0
 [3.6.0]: https://github.com/giantswarm/app/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/giantswarm/app/compare/v3.4.0...v3.5.0
 [3.4.0]: https://github.com/giantswarm/app/compare/v3.3.0...v3.4.0